% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senanal.R
\name{senanal}
\alias{senanal}
\title{One-way sensitivity analysis for CEA}
\usage{
senanal(data, unit)
}
\arguments{
\item{data}{A data frame with four columns: 1) parameters 2) base value 3) lower bound and 4) upper bound.}

\item{unit}{The unit of the outcome variable, e.g. 2022 USD.}
}
\value{
A tornado diagram and a result data frame for one-way sensitivity analysis
}
\description{
This is a function that provides an automatic calculation for an one-way sensitivity analysis used in CEA.
Make sure that install three packages before loading this package: `ggplot2`, `plyr`, `dplyr` and `tidyverse`.
}
\examples{
mydata = data.frame(cbind("Parameter" = c("Faculty", "Student", "Doctor"),
                          "Base" = c(2000, 7200, 3000),
                          "Low" = c(1000,6000,2000),
                          "High" = c(3000, 9000, 5000)))
senanal(data = mydata, unit = "USD")

}
